--- test/render/test-cases/line-layer.js
@@ -1,34 +1,0 @@
-// deck.gl
-// SPDX-License-Identifier: MIT
-// Copyright (c) vis.gl contributors
-
-/* eslint-disable callback-return */
-import {LineLayer} from '@deck.gl/layers';
-import {routes} from 'deck.gl-test/data';
-
-export default [
-  {
-    name: 'line-lnglat',
-    viewState: {
-      latitude: 37.751537058389985,
-      longitude: -122.42694203247012,
-      zoom: 11.5,
-      pitch: 0,
-      bearing: 0
-    },
-    layers: [
-      new LineLayer({
-        id: 'line-lnglat',
-        data: routes,
-        opacity: 0.8,
-        getWidth: 0,
-        widthMinPixels: 2,
-        getSourcePosition: d => d.START,
-        getTargetPosition: d => d.END,
-        getColor: d => (d.SERVICE === 'WEEKDAY' ? [255, 64, 0] : [255, 200, 0]),
-        pickable: true
-      })
-    ],
-    goldenImage: './test/render/golden-images/line-lnglat.png'
-  }
-];
